Refactor db/models.py and src/schemas.py
file with pydantic models renamed to schemas.py not to be confused with 2 models.py files

all db models moved to one file db/models.py as it makes all cleaner
models and schemas are reformatted
utils.py reformatted according to new models and schemas
