+++ src/schemas.py
@@ -1,5 +1,3 @@
-from datetime import datetime
-
 from pydantic import BaseModel
 
 
@@ -7,22 +5,19 @@ class UserModel(BaseModel):
     """User object model"""
 
     name: str
-    user_id: str
 
 
 class MessageModel(BaseModel):
     """Message object model"""
 
-    message_id: int
-    user_id: str
-    body: str
-    timestamp: datetime
+    user: str
+    text: str
 
 
 class ChatRoomModel(BaseModel):
     """ChatRoom object model"""
 
-    chatroom_id: str
+    name: str
     messages: list[MessageModel]
     users: list[UserModel]
 
@@ -45,5 +40,5 @@ class SendMessageResponseModel(BaseModel):
 class ChatRoomResponseModel(BaseModel):
     """Get Messages Response Model"""
 
-    chatroom_id: str
+    name: str
     messages: list[MessageModel]
