+++ src/tests/test_socket.py
@@ -2,7 +2,7 @@ import json
 
 import pytest
 
-from src.models import SignUpResponseModel, UserModel
+from src.schemas import SignUpResponseModel, UserModel
 from src.sockets.connection_handler import ConnectionHandler
 from src.tests.mocks.socket_mock import SocketMock
